+++ src/components/tour/index.ts
@@ -1,0 +1,8 @@
+export { EuiTour, EuiTourProps } from './tour';
+export { EuiTourStep, EuiTourStepProps } from './tour_step';
+export {
+  EuiTourStepIndicator,
+  EuiTourStepIndicatorProps,
+} from './tour_step_indicator';
+export { useEuiTour, EuiStatelessTourStep } from './useEuiTour';
+export * from './types';